requirements.txt
Flask==1.0.2
google-cloud-datastore==1.7.1
google-api-python-client==1.7.9
